pyproject.toml
[project]
name = "cvmaker"
version = "0.1.0"
description = "A CV composer. Write each experience once, translate it to the languages you care about, then assemble the right CV for each role at the click of a button."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Natan Sisoev", email = "natan.sisoev@gmail.com" },
]
license = { text = "MIT" }
keywords = ["django", "cv", "resume", "typst", "rendercv"]
classifiers = [
    "Framework :: Django :: 5.1",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]

# -------------------------------------------------------------------
# Runtime dependencies
# -------------------------------------------------------------------
# Grouped; each group has a reason. If you're adding a new dep, add it
# to the right group and pin the minor version. We'll lock the exact
# versions in uv.lock.
# -------------------------------------------------------------------
dependencies = [
    # --- Django + extensions ---
    "Django>=5.1,<5.2",
    "django-crispy-forms>=2.3,<3",       # Phase 4 replaces with django-cotton components
    "django-environ>=0.11,<1",
    "django-model-utils>=5.0,<6",         # InheritanceManager for polymorphic entries
    "django-widget-tweaks>=1.5,<2",       # Phase 4 replaces with Tailwind forms

    # --- Database ---
    "dj-database-url>=2.3,<3",
    "psycopg[binary]>=3.2,<4",            # psycopg3 — native async, supersedes psycopg2

    # --- Server ---
    "gunicorn>=23,<24",
    "uvicorn[standard]>=0.34,<1",         # ASGI for async + HTMX
    "whitenoise>=6.9,<7",                 # static serving (Phase 8 may move to CDN)

    # --- Rendering: rendercv drives Typst via subprocess ---
    "rendercv>=2.2,<3",
    "rendercv-fonts>=0.4,<1",

    # --- YAML parsing ---
    # Used directly in apps/cv/models.py as a stand-in for rendercv 1.x's
    # `read_a_yaml_file`, which was removed in rendercv 2.x. Already a
    # transitive dep via rendercv; pinned here so the direct use is explicit.
    "pyyaml>=6.0,<7",

    # --- Images ---
    # Required by Django's ImageField (apps/cv/models.py: CVInfo.photo).
    # Django's system check hints but doesn't install it for us.
    "pillow>=11,<12",
]

# -------------------------------------------------------------------
# Dev + test dependencies
# -------------------------------------------------------------------
[dependency-groups]
dev = [
    # linting / formatting / types
    "ruff>=0.8,<1",
    "djlint>=1.36,<2",
    "mypy>=1.13,<2",
    "django-stubs[compatible-mypy]>=5.1,<6",
    "django-upgrade>=1.22,<2",
    "pre-commit>=4,<5",

    # testing
    "pytest>=8,<9",
    "pytest-django>=4.9,<5",
    "pytest-cov>=6,<7",
    "pytest-xdist>=3.6,<4",
    "factory-boy>=3.3,<4",
    "faker>=33,<34",
    "pytest-playwright>=0.5,<1",          # Phase 1.6 e2e smoke tests

    # dev-only convenience
    "django-debug-toolbar>=4.4,<5",
    "django-extensions>=3.2,<4",
    "ipython>=8.30,<9",
    "werkzeug>=3.1,<4",                   # pulled in by runserver_plus
]

# -------------------------------------------------------------------
# uv
# -------------------------------------------------------------------
[tool.uv]
package = false                              # this is an app, not a library — don't build wheels
default-groups = ["dev"]                     # `uv sync` installs dev by default; `uv sync --no-dev` for prod
managed = true

# -------------------------------------------------------------------
# Ruff — lint + format (replaces black, isort, flake8, pyupgrade)
# -------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "apps", "tests"]
extend-exclude = ["migrations", ".venv", "node_modules", "static/dist"]

[tool.ruff.lint]
select = [
    "E", "F", "W",      # pycodestyle + pyflakes
    "I",                 # isort
    "UP",                # pyupgrade
    "B",                 # bugbear
    "C4",                # comprehensions
    "DJ",                # flake8-django
    "SIM",               # simplify
    "RUF",               # ruff-specific
    "T20",               # no print()
    "TID",               # tidy imports
]
ignore = [
    "E501",              # line length enforced by the formatter
    "DJ001",              # nullable CharField — we'll opt in per-field with noqa
]

[tool.ruff.lint.isort]
known-first-party = ["cvmaker", "accounts", "cv", "entries", "sections", "core", "rendering"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]                        # asserts are fine in tests
"**/migrations/*" = ["E501", "UP"]           # autogenerated

[tool.ruff.format]
quote-style = "double"
line-ending = "lf"

# -------------------------------------------------------------------
# djlint — Django template linting
# -------------------------------------------------------------------
[tool.djlint]
profile = "django"
indent = 2
max_line_length = 120
format_css = true
format_js = true

# -------------------------------------------------------------------
# mypy — gradual typing, strict where possible
# -------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
mypy_path = ["src", "apps"]
plugins = ["mypy_django_plugin.main"]
exclude = ["migrations/", ".venv/", "node_modules/"]
ignore_missing_imports = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["*.migrations.*"]
ignore_errors = true

[tool.django-stubs]
django_settings_module = "cvmaker.settings.dev"

# -------------------------------------------------------------------
# pytest
# -------------------------------------------------------------------
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "cvmaker.settings.test"
python_files = ["test_*.py", "*_test.py", "tests.py"]
testpaths = ["tests", "apps"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=apps",
    "--cov=src/cvmaker",
    "--cov-report=term-missing:skip-covered",
    "--cov-fail-under=0",                    # Phase 1.6 ratchets this to 70
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:rendercv.*",
]
markers = [
    "slow: marks tests that take > 1s (deselect with '-m \"not slow\"')",
    "e2e: end-to-end browser tests",
]

# -------------------------------------------------------------------
# Coverage
# -------------------------------------------------------------------
[tool.coverage.run]
branch = true
source = ["apps", "src/cvmaker"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "manage.py",
    "*/wsgi.py",
    "*/asgi.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
